+++ src/fife/view/renderers/genericrenderer.h
@@ -8,6 +8,7 @@
 #include "platform.h"
 
 // Standard C++ library includes
+#include <cstdint>
 #include <map>
 #include <string>
 #include <vector>
@@ -156,7 +157,7 @@ namespace FIFE
         private:
             RendererNode m_anchor;
             AnimationPtr m_animation;
-            uint32_t m_start_time;
+            uint64_t m_start_time;
             float m_time_scale;
             bool m_zoomed;
     };
